requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
mutagen>=1.47.0
sqlalchemy>=2.0.36
pydantic>=2.11.0
pyyaml>=6.0.2
aiofiles>=24.1.0
watchfiles>=1.0.0
httpx>=0.28.0
jinja2>=3.1.5
pyacoustid>=1.2.0
musicbrainzngs>=0.7.1
httpx[socks]
